Agile Scrum Team Projects/Congo Online/server/src/test/java/GameTest.java
import Game.Game;
import webconnection.Action;
import database.DatabaseHandler;
import org.junit.Before;
import org.junit.Test;
import org.junit.After;
import java.sql.SQLException;
import static org.junit.Assert.*;

/**
Commenting out tests for the sake of Travis CI
*/
public class GameTest {

    Game game;
   
    @Before
    public void initialize() {
        game = new Game();
    }
   
   //@Test
   public void testLoadExistingGame() throws Exception {
        Action action = new Action();
        action.communicationType = "requestMoves";
        action.matchID = "1";
        
        Game game = new Game();
        game.loadExistingGame(action);
   }

    @Test
    public void testMoveSequenceCorrectPlayer1() throws Exception {
        Action action = new Action();
        action.playerOneName = "Sue";
        action.userName = "Sue";
        game.setActivePlayer("Sue");

        game.getGameBoard().initialize();
        int location = 16;  //player 1 CAN move pawn initially in second row
        assertTrue(game.moveSequenceCorrect(action, game, location));
    }

    @Test
    public void testMoveSequenceCorrectPlayer2() throws Exception {
        Action action = new Action();
        action.playerOneName = "Sue";
        action.userName = "Joe";
        game.setActivePlayer("Joe");

        game.getGameBoard().initialize();
        int location = 56;  //player 2 CAN move pawn initially in fifth row
        assertTrue(game.moveSequenceCorrect(action, game, location));
    }

    @Test(expected = Exception.class)
    public void testMoveSequenceCorrectNullPiece() throws Exception {
        Action action = new Action();
        action.playerOneName = "Sue";
        action.userName = "Joe";
        game.setActivePlayer("Joe");

        game.getGameBoard().initialize();
        int location = 33;  //no player can move a piece that isn't there

        /* try to move non-existent piece to see if exception is thrown */
        game.moveSequenceCorrect(action, game, location);

    }

    @Test
    public void testMoveSequenceCorrectWrongSeq() throws Exception {
        Action action = new Action();
        action.playerOneName = "Sue";
        action.userName = "Sue";
        game.setActivePlayer("Joe");

        game.getGameBoard().initialize();
        int location = 14;  //this is player's piece
        assertTrue(game.moveSequenceCorrect(action, game, location) == false);
    }
}
